separate embeddings from mongodb, to allow for more than openai embeddings
