+++ src/memorizz/database/mongodb/__init__.py
@@ -1,3 +1,3 @@
-from .mongodb_tools import MongoDBTools, MongoDBToolsConfig, get_embedding, get_mongodb_toolbox
+from .mongodb_tools import MongoDBTools, MongoDBToolsConfig, get_mongodb_toolbox
 
-__all__ = ['MongoDBTools', 'MongoDBToolsConfig', 'get_embedding', 'get_mongodb_toolbox']
+__all__ = ['MongoDBTools', 'MongoDBToolsConfig', 'get_mongodb_toolbox']
